Creating and HTML response
We rarely have to deal with responses from web servers in plain text.
We need to extend the previous code so that the server responds with a response in HTML just as a normal web server would.

Adjusted the content type which is now text/html
We dropped the response.write() method on line 6 resulting all ythe HTTP body to be sent entirely using the end method.